chore: Flutter upgrade

+++ github_pages_builder.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+flutter build web --release
+# Check if the build was successful
+if [ $? -ne 0 ]; then
+  echo "Flutter build failed. Exiting."
+  exit 1
+fi
+
 # Replace "<base href="$FLUTTER_BASE_HREF">" with "<base href="./">" in web/index.html
 sed -i.bak 's#<base href="\$FLUTTER_BASE_HREF">#<base href="./">#g' web/index.html
 rm -rf web/index.html.bak
